refactor: take out the processing of the routes
